Update SystemRole & Permission

+++ routes/todoRoutes.js
@@ -3,7 +3,8 @@ const router = express.Router();
 const { PrismaClient } = require("@prisma/client");
 const prisma = new PrismaClient();
 const { body, validationResult } = require("express-validator");
-const auth = require("../middleware/auth.js");
+const { auth, isAdmin } = require("../middleware/auth.js");
+const checkPermission = require("../middleware/checkPermission.js");
 
 // Route pour obtenir les todos de l'utilisateur connecté
 router.get("/", auth, async (req, res) => {
@@ -26,9 +27,7 @@ router.get("/", auth, async (req, res) => {
 router.post(
   "/create",
   auth,
-  [
-    body("title").notEmpty().withMessage("Le titre est obligatoire."),
-  ],
+  [body("title").notEmpty().withMessage("Le titre est obligatoire.")],
   async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -56,30 +55,34 @@ router.post(
 );
 
 // Route DELETE pour supprimer un todo
-router.delete("/delete/:id", auth, async (req, res) => {
-  try {
-    const todo = await prisma.todo.delete({
-      where: {
-        id: parseInt(req.params.id),
-        userId: req.userId,
-      },
-    });
-    res.json({ message: "Todo supprimé avec succès" });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({
-      error: "Une erreur est survenue lors de la suppression du todo.",
-    });
+router.delete(
+  "/delete/:id", 
+  auth, 
+  isAdmin, 
+  checkPermission('delete:todos'),
+  async (req, res) => {
+    try {
+      const todo = await prisma.todo.delete({
+        where: {
+          id: parseInt(req.params.id),
+          userId: req.userId,
+        },
+      });
+      res.json({ message: "Todo supprimé avec succès" });
+    } catch (error) {
+      console.error(error);
+      res.status(500).json({
+        error: "Une erreur est survenue lors de la suppression du todo.",
+      });
+    }
   }
-});
+);
 
 // Route PUT pour mettre à jour un todo
 router.put(
   "/update/:id",
   auth,
-  [
-    body("title").notEmpty().withMessage("Le titre est obligatoire."),
-  ],
+  [body("title").notEmpty().withMessage("Le titre est obligatoire.")],
   async (req, res) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
